+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.19.0 (provisional)
 
 * Adding the `clustering/record-linkage/leader` namespace.
+* Adding the `keyers/name-power-set` namespace.
 * Adding the `keyers/normalize` namespace.
 * Adding the `tokenizers/skipgrams` namespace.
 * Fixing & optimizing `clustering/k-means`.